Split `JoeClass` into a `JoeClass` interface and a `ScriptedClass` implementation.
